Implemented the first part of story selection
There are a lot of other changes because I wasn't sure how to pass information from child components to parent components. I spent a lot of time trying to figure out the React way to do this so were lots of other minor change that I made while trying to figure this out. I'm leaving the lifecycle functions in because I found it very useful to see when they are being called. There is still more to do with the story selection, but this code is really messy and hard to understand so I think I need to do some serious clean-up before proceeding.