Changed the when the stories are grabbed
I made this change because it takes a long time to get all the stories back, so the initial page load is really slow. This way, at least the header and footer are rendered immediately. I intend to do more optimization with this soon.

I made is so that stories are updated every ten minutes. :)