Continued with comment rendering
Made it so that all comments are displayed, not just direct children of the story

Made is so that HTML embedded in comments is displayed as HTML instead of as text. (Fingers crossed that HN won't send me dangerous HTML!)

Started fixing up comment styles, mostly just wanted to get the offset that depends on the level of nesting working. :)