Correctly formatted the times
There seems to be an occasional difference between the time I am displaying and the time displayed on the actual HN site. I checked my date formatting and I even tried integrating Moment.js, but I still got the same error, so I don't know... maybe HN is giving me the wrong date on some of these? Or else maybe I am somehow messing up the date before I get to the point of fomatting it. I don't notice a pattern in the incorrect dates, so for now I'm just going to make a note of this.